Add interpolate module
